feat: created a select type component
